as of 28 feb 2023
we can move forward with the final architecture of the game. the game runs stable with 5 players + 2 client and server. with 4 players it runs perfectly on my PC.

Time testing shows that navigation part of the ship consumes most time.

overall processing is taking more time. sending and receiving are taking negligible time and are almost perfect.